chore: split main
